+++ azuredevops/internal/service/taskagent/resource_agentpool.go
@@ -1,4 +1,4 @@
-package azuredevops
+package taskagent
 
 import (
 	"fmt"
@@ -7,14 +7,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/microsoft/azure-devops-go-api/azuredevops/taskagent"
-	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/client"
+	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/converter"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/suppress"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/validate"
 )
 
-func resourceAzureAgentPool() *schema.Resource {
+func ResourceAgentPool() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceAzureAgentPoolCreate,
 		Read:   resourceAzureAgentPoolRead,
